feat: rollback to TCP client
