+++ docker-compose.yaml
@@ -22,6 +22,8 @@ services:
       context: .
       dockerfile: ./server/Dockerfile
     container_name: server
+    ports:
+      - "8081:8080"  # TCP port for client connections
     depends_on:
       rabbitmq:
         condition: service_healthy
@@ -44,6 +46,7 @@ services:
       streaming-service:
         condition: service_started
     environment:
+      - SERVER_PORT=8080
       - RABBITMQ_HOST=rabbitmq
       - RABBITMQ_PORT=5672
       - RABBITMQ_USER=admin
@@ -56,17 +59,13 @@ services:
       dockerfile: ./client/Dockerfile
     container_name: client
     depends_on:
-      rabbitmq:
-        condition: service_healthy
+      server:
+        condition: service_started
     volumes:
       - ./data:/app/data
     environment:
       - CLIENT_ID=CLI1
-      - RABBITMQ_HOST=rabbitmq
-      - RABBITMQ_PORT=5672
-      - RABBITMQ_USER=admin
-      - RABBITMQ_PASS=password
-    command: ["./main", "/app/data"]
+    command: ["./main", "/app/data", "server:8080"]
     profiles: ["data-flow"]
 
   client-2:
@@ -75,17 +74,13 @@ services:
       dockerfile: ./client/Dockerfile
     container_name: client-2
     depends_on:
-      rabbitmq:
-        condition: service_healthy
+      server:
+        condition: service_started
     volumes:
       - ./data:/app/data
     environment:
       - CLIENT_ID=CLI2
-      - RABBITMQ_HOST=rabbitmq
-      - RABBITMQ_PORT=5672
-      - RABBITMQ_USER=admin
-      - RABBITMQ_PASS=password
-    command: ["./main", "/app/data"]
+    command: ["./main", "/app/data", "server:8080"]
     profiles: ["data-flow"]
 
   # Worker services
